moved managerpage2 to pages folder
