+++ client/src/App.js
@@ -13,7 +13,7 @@ import './App.css';
 
 // new manager components
 import ManagerPageTwo from "./pages/ManagerPageTwo"
-import ManagerMenuItemsPage from "./managerStyle/ManagerMenuItemsPage"
+import ManagerMenuItemsPage from "./pages/ManagerMenuItemsPage"
 import newNavBar from "./components/Waitstaff/newNavBar"
 
 class App extends Component {
@@ -55,7 +55,8 @@ class App extends Component {
           {/* <Guest1SignUpLogin /> */}
           {/* <Guest2ResMake /> */}
           {/* <Guest3ResConf /> */}
-          {/* <ManagerPage /> */}
+          {/* <ManagerPageTwo /> */}
+          {/* <ManagerMenuItemsPage /> */}
           {/* <Waitstaff1CurTablesPage /> */}
           <Waitstaff2OrderPage menu={this.state.menu} />
           {/* <Waitstaff3ReceiptPage /> */}
