+++ client/src/pages/ManagerMenuItemsPage.js
