+++ client/src/pages/navBarTest.js
@@ -1,0 +1,13 @@
+import React from "react";
+
+
+function navBarTest(){
+  return (
+  <div>
+    
+  </div>
+  )
+}
+
+
+export default navBarTest;
